requirements.txt
# OpenAI API（音声認識・AI応答・TTS）
openai>=1.0.0

# 音声処理
pyaudio>=0.2.11

# WebSocket クライアント
websockets>=11.0

# 設定管理
python-dotenv>=1.0.0

# Google Sheets連携
gspread>=5.0.0
google-auth>=2.0.0

# メール通知
# smtplib は標準ライブラリのため不要

# 数学・データ処理（標準ライブラリのmathを使用）